Add Honeydew example app.
Consider adding this example of the CELP stack. Commanded v1.3 and LV v1.16.
Domain is a TODO list (or Honey Do list... Get it 😄). Some part's left unfinished for the user to complete (simple and example given in code).
The idea is to lower the barrier for devs new to Commanded/CQRS/ES in general by giving them a playground that's already setup.